--- a/crates/edr_chain_l1/src/hardfork.rs
+++ b/crates/edr_chain_l1/src/hardfork.rs
@@ -1,0 +1,3 @@
+//! Ethereum L1 hardfork-related types.
+
+pub use revm_primitives::hardfork::name;