--- a/crates/edr_chain_l1/src/lib.rs
+++ b/crates/edr_chain_l1/src/lib.rs
@@ -1,0 +1,30 @@
+//! Ethereum L1 chain types
+
+mod eip2718;
+pub mod pooled;
+pub mod request;
+pub mod signed;
+mod spec;
+pub mod r#type;
+
+use edr_evm_spec::EvmSpecId;
+pub use revm_context_interface::result::OutOfGasError;
+
+pub use self::{
+    eip2718::TypedEnvelope, pooled::PooledTransaction, r#type::Type, request::Request,
+    signed::Signed,
+};
+
+pub type BlockEnv = revm_context::BlockEnv;
+
+/// Ethereum L1 halt reason.
+pub type HaltReason = revm_context_interface::result::HaltReason;
+
+/// Ethereum L1 hardfork.
+pub type Hardfork = EvmSpecId;
+
+/// Ethereum L1 invalid header error.
+pub type InvalidHeader = revm_context_interface::result::InvalidHeader;
+
+/// Ethereum L1 invalid transaction error.
+pub type InvalidTransaction = revm_context_interface::result::InvalidTransaction;