--- a/crates/edr_chain_l1/src/lib.rs
+++ b/crates/edr_chain_l1/src/lib.rs
@@ -1,6 +1,7 @@
 //! Ethereum L1 chain types
 
 mod eip2718;
+pub mod hardfork;
 pub mod pooled;
 pub mod request;
 pub mod signed;
@@ -8,13 +9,16 @@ mod spec;
 pub mod r#type;
 
 use edr_evm_spec::EvmSpecId;
+pub use revm_context::TxEnv;
 pub use revm_context_interface::result::OutOfGasError;
+pub use revm_primitives::hardfork::UnknownHardfork;
 
 pub use self::{
     eip2718::TypedEnvelope, pooled::PooledTransaction, r#type::Type, request::Request,
-    signed::Signed,
+    signed::Signed, spec::L1ChainSpec,
 };
 
+/// Ethereum L1 block environment.
 pub type BlockEnv = revm_context::BlockEnv;
 
 /// Ethereum L1 halt reason.
@@ -28,3 +32,6 @@ pub type InvalidHeader = revm_context_interface::result::InvalidHeader;
 
 /// Ethereum L1 invalid transaction error.
 pub type InvalidTransaction = revm_context_interface::result::InvalidTransaction;
+
+/// L1 Ethereum chain type
+pub const CHAIN_TYPE: &str = "L1";