--- a/crates/edr_eip2930/src/lib.rs
+++ b/crates/edr_eip2930/src/lib.rs
@@ -1,0 +1,3 @@
+//! Types related to EIP-2930.
+
+pub use revm_context_interface::transaction::{AccessList, AccessListItem};