--- a/crates/edr_eip7702/src/lib.rs
+++ b/crates/edr_eip7702/src/lib.rs
@@ -1,3 +1,5 @@
+//! Types related to EIP-7702.
+
 pub use revm_context_interface::transaction::{Authorization, SignedAuthorization};
 /// Re-export EIP-7702 constants.
 pub use revm_primitives::eip7702::{PER_AUTH_BASE_COST, PER_EMPTY_ACCOUNT_COST};