--- a/crates/edr_eth/src/eips.rs
+++ b/crates/edr_eth/src/eips.rs
@@ -1,5 +1,3 @@
-/// Types related to EIP-2718.
-pub mod eip2718;
 /// Types related to EIP-4844.
 pub mod eip4844;