--- a/crates/edr_eth/src/eips/eip2930.rs
+++ b/crates/edr_eth/src/eips/eip2930.rs
@@ -1,1 +1,0 @@
-pub use revm_context_interface::transaction::{AccessList, AccessListItem};