This PR serves to prove that it's possible to move the `L1ChainSpec` into its own crate, instead of its implementation being part of the "core" EDR crates (`edr_eth`, `edr_evm`, `edr_provider`, `edr_napi_core`, `edr_napi`).

A prior attempt (https://github.com/NomicFoundation/edr/pull/1006) failed because the tests of the "core" EDR crates depend on `L1ChainSpec` - and its defined types - causing a recursive dependency.

This PR worked around this limitation by introducing a new shared dependency called `edr_evm_spec` containing the `ChainHardfork`, `ChainSpec`, and `EthHeaderConstants` trait. This allows `edr_chain_l1` to depend on `edr_evm_spec` to implement the traits, instead of depending on `edr_eth`; breaking the recursive dependency.

This change required splitting `edr_eth` into several smaller crates to ensure there were no recursive dependencies.

To fully separate the `L1ChainSpec` implementation, follow-up PRs need to repeat this pattern for:
- [ ] `edr_rpc_eth::RpcSpec`
- [ ] `edr_evm::RuntimeSpec`
- [ ] `edr_provider::ProviderSpec`
- [ ] `edr_napi::SyncNapiSpec`

## Benchmarks

One potential side-effect of moving code into many, smaller crates is improved incremental compilation times. I ran some benchmarks to compare compile times for common development workflows:

```bash
cargo clippy --all-features --all-targets  
```

| Test Run | Main (8a48e24) | PR #1067 |  Main -> PR #1067 |
|----------|------|-----|-----|
| clean | 677.75s user<br/>83.24s system<br/>55.020s total | 675.57s user<br/>84.18s system<br/>55.419s total | -0.32% user<br/>+1.13% system<br/>+0.73% total |
| edit `edr_eth/src/reward_percentile.rs` | 25.26s user<br/>3.57s system<br/>10.598s total | 24.99s user<br/>3.93s system<br/>10.227s total | -1.07% user<br/>+10.08% system<br/>-3.50% total |
| edit `edr_eth/src/l1.rs` OR<br/>edit `edr_chain_l1/src/lib.rs` | 24.63s user<br/>3.71s system<br/>10.412s total | 23.67s user<br/>3.54s system<br/>9.750s total | -3.90% user<br/>-4.58% system<br/>-6.36% total |
| edit `edr_rpc_eth/src/override.rs` | 24.78s user<br/>3.35s system<br/>9.958s total | 23.69s user<br/>3.53s system<br/>9.738s total | -4.40% user<br/>+5.37% system<br/>-2.21% total |
| edit `edr_evm/src/mempool.rs` | 23.14s user<br/>3.60s system<br/>9.312s total | 22.63s user<br/>3.28s system<br/>9.159s total | -2.20% user<br/>-8.89% system<br/>-1.64% total |
| edit `edr_provider/src/data.rs` | 8.85s user<br/>1.46s system<br/>3.918s total | 8.71s user<br/>1.55s system<br/>3.875s total | -1.58% user<br/>+6.16% system<br/>-1.10% total |
| edit `edr_napi_core/src/logger.rs` | 4.72s user<br/>0.84s system<br/>2.481s total | 4.49s user<br/>1.09s system<br/>2.522s total | -4.87% user<br/>+29.76% system<br/>+1.65% total |
| edit `edr_op/src/lib.rs` | 4.14s user<br/>0.71s system<br/>2.269s total | 4.18s user<br/>0.71s system<br/>2.290s total | +0.97% user<br/>0.00% system<br/>+0.93% total |
| edit `edr_napi/src/cast.rs` | 2.70s user<br/>0.55s system<br/>1.889s total | 2.88s user<br/>0.41s system<br/>1.940s total | +6.67% user<br/>-25.45% system<br/>+2.70% total |

```bash
cd crates/edr_napi && pnpm build
```

| Test Run | Main (8a48e24) | PR #1067 | Main -> PR #1067 |
|----------|------|-----|-----|
| clean | 1171.45s user<br/>79.30s system<br/>320.48s total | 1112.96s user<br/>75.39s system<br/>308.48s total | -4.99% user<br/>-4.93% system<br/>-3.74% total |

```bash
cd crates/edr_napi && pnpm build:dev
```

| Test Run | Main (8a48e24) | PR #1067 |  Main -> PR #1067 |
|----------|------|-----|-----|
| clean | 1670.73s user<br/>94.93s system<br/>118.67s total | 1686.80s user<br/>94.81s system<br/>119.58s total | +0.96% user<br/>-0.13% system<br/>+0.77% total |
| edit `edr_eth/src/reward_percentile.rs` | 388.34s user<br/>6.57s system<br/>54.172s total | 391.45s user<br/>6.15s system<br/>53.000s total | +0.80% user<br/>-6.39% system<br/>-2.16% total |
| edit `edr_eth/src/l1.rs` OR<br/>edit `edr_chain_l1/src/lib.rs` | 387.71s user<br/>6.27s system<br/>54.005s total | 402.74s user<br/>6.11s system<br/>54.129s total | +3.88% user<br/>-2.55% system<br/>+0.23% total |
| edit `edr_rpc_eth/src/override.rs` | 386.27s user<br/>6.26s system<br/>54.088s total | 388.89s user<br/>6.11s system<br/>52.364s total | +0.68% user<br/>-2.40% system<br/>-3.19% total |
| edit `edr_evm/src/mempool.rs` | 388.19s user<br/>6.42s system<br/>53.429s total | 388.22s user<br/>6.50s system<br/>51.715s total | +0.01% user<br/>+1.25% system<br/>-3.21% total |
| edit `edr_provider/src/data.rs` | 286.93s user<br/>4.40s system<br/>38.206s total | 291.09s user<br/>3.83s system<br/>36.922s total | +1.45% user<br/>-12.95% system<br/>-3.36% total |
| edit `edr_napi_core/src/logger.rs` | 289.00s user<br/>3.86s system<br/>36.162s total | 283.55s user<br/>3.82s system<br/>34.194s total | -1.89% user<br/>-1.04% system<br/>-5.44% total |
| edit `edr_op/src/lib.rs` | 282.69s user<br/>3.88s system<br/>35.391s total | 293.15s user<br/>4.00s system<br/>34.681s total | +3.70% user<br/>+3.09% system<br/>-2.01% total |
| edit `edr_napi/src/cast.rs` | 278.80s user<br/>3.72s system<br/>34.569s total | 287.96s user<br/>3.49s system<br/>33.540s total | +3.29% user<br/>-6.18% system<br/>-2.98% total |