@Wodann 2025-09-02
> It seems that we're having CI timeouts again, but I don't think they can be caused by these changes. From what I'm seeing in the logs, [this test](https://github.com/NomicFoundation/edr/blob/8a48e24bf9c86fb84998aa600fd6c5a0544a398d/hardhat-tests/test/internal/hardhat-network/provider/logs.ts#L71) is the one that times out.

I agree, although I fear this might become a recurring issue due to (relatively) recent changes in GitHub-hosted macOS runners. I started a thread in our dev channel about this.