@Wodann 2025-09-02
> One question for my knowledge, is there a good practice rule when to fully qualify vs fully import paths? I think I've seen different approaches in our codebase and I was wondering what's recommended. Should we only fully qualify paths that are name conflicts?

I've been evolving my own best practice. I know some users of APIs prefer `path::to::fully::named::Type` whereas other users prefer `ImportedFullyNamedType`. I tried to add support for users that prefer the former method but ran into the issue that having many `Type`s that have a similar name makes IDE features for looking for that type (`Ctrl + T` in VSCode) less useful.

My next iteration would be one where we combine the two approaches. E.g. we'd have:

```rs
pub type Signed = L1SignedTransaction;
// OR: pub use Signed = L1SignedTransaction;

pub enum L1SignedTransaction { .. }
```